+++ src/InputFileCollector.php
@@ -6,7 +6,7 @@ namespace Vjik\TelegramBot\Api;
 
 use Vjik\TelegramBot\Api\Type\InputFile;
 
-final class RequestFileCollector
+final class InputFileCollector
 {
     private int $counter;
 
@@ -16,7 +16,7 @@ final class RequestFileCollector
     private array $files = [];
 
     public function __construct(
-        private string $prefix = 'file',
+        private readonly string $prefix = 'file',
         int $start = 0,
     ) {
         $this->counter = $start;