+++ src/Type/InputMediaPhoto.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Vjik\TelegramBot\Api\Type;
 
-use Vjik\TelegramBot\Api\RequestFileCollector;
+use Vjik\TelegramBot\Api\InputFileCollector;
 
 /**
  * @see https://core.telegram.org/bots/api#inputmediaphoto
@@ -28,7 +28,7 @@ final readonly class InputMediaPhoto implements InputMedia
         return 'photo';
     }
 
-    public function toRequestArray(?RequestFileCollector $fileCollector = null): array
+    public function toRequestArray(?InputFileCollector $fileCollector = null): array
     {
         if ($fileCollector !== null) {
             $media = $this->media instanceof InputFile