+++ src/Type/InputPaidMedia.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Vjik\TelegramBot\Api\Type;
 
-use Vjik\TelegramBot\Api\RequestFileCollector;
+use Vjik\TelegramBot\Api\InputFileCollector;
 
 /**
  * @see https://core.telegram.org/bots/api#inputpaidmedia
@@ -13,5 +13,5 @@ interface InputPaidMedia
 {
     public function getType(): string;
 
-    public function toRequestArray(?RequestFileCollector $fileCollector = null): array;
+    public function toRequestArray(?InputFileCollector $fileCollector = null): array;
 }