+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.6.0 under development
 
 - Chg #132: Rename definition "telegram client" to "transport". Classes, namespaces, and variables rename accordingly.
+- Chg #132: Rename `RequestFileCollector` to `InputFileCollector` and move to root namespace.
 
 ## 0.5.0 January 02, 2025